+++ lib/pubnub/validators/common_validator.rb
@@ -52,6 +52,22 @@ module Pubnub
           )
         end
       end
+
+      def validate_user_id(user_id)
+        if !user_id || user_id.blank?
+          raise(
+            InitializationError.new,
+            'Missing required :user_id parameter.'
+          )
+        elsif !([String, Symbol].include?(user_id.class) ||
+          user_id.blank?)
+          raise(
+            InitializationError.new,
+            'user_id parameter is not valid. \
+                   Should be type of String or Symbol.'
+          )
+        end
+      end
     end
   end
 end