+++ spec/lib/connection_callback_spec.rb
@@ -26,7 +26,7 @@ describe "Connect, disconnect and reconnect callbacks" do
       end,
     )
 
-    @pubnub = Pubnub.new(
+    @pubnub = Pubnub.new(user_id: Pubnub::UUID.generate, 
       subscribe_key: :demo,
       publish_key: :demo,
       reconnect_interval: 0,