+++ spec/lib/keep_alive_spec.rb
@@ -1,9 +1,0 @@
-require "spec_helper"
-
-describe "keep_alive connection" do
-  around :each do |example|
-    example.run_with_retry retry: 10
-  end
-
-  it ""
-end